backend/requirements.txt
alembic==1.8.0
aniso8601==9.0.1
anyio==3.6.1
attrs==21.4.0
automium==0.2.6
automium-web==0.1.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
clients==1.3
colorama==0.4.5
config==0.5.1
cryptography==37.0.4
data==0.4
decorator==5.1.1
Flask==2.1.2
Flask-Cors==3.0.10
flask-expects-json==1.7.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restful-swagger-3==0.5.1
Flask-SQLAlchemy==2.5.1
funcsigs==1.0.2
functions==0.7.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
legacy==0.1.6
ma==1.0.2
ma.mongo==1.0.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
rfc3986==1.5.0
schemas==0.7.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
tomli==2.0.1
urllib3==1.26.11
Werkzeug==2.1.2
